Delete 'docs' row when a document is purged/expired
Until now, rows in the 'docs' table were never removed. This becomes
more of a problem now that large numbers of docs might get created and
expired, causing the table to grow monotonically.

Fixing this was hard to do, because it invalidates the docID-to-row
cache (_docIDs), and Cache Invalidation Is Hard. Solution was to:

* Expose the read-lock from FMDatabase
* Access the cache (and use the resulting row#) only while holding the
  database's read lock, or in a transaction (which also holds the lock)
* Share the cache between all storage instances open on the same db path

For #1181